requirements.txt
streamlit
tensorflow-cpu==2.13.0
numpy
Pillow